# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/authrequired/teacher": {
            "patch": {
                "description": "修改教师相关信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "修改教师相关信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "查询过滤器，如果没有默认查询所以信息",
                        "name": "action",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "修改邮箱",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "修改真实姓名",
                        "name": "realName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "修改学号",
                        "name": "userSchoollD",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "修改学校代码",
                        "name": "schoolCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "修改班级",
                        "name": "class",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "修改性别",
                        "name": "sex",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/teacher/password": {
            "patch": {
                "description": "修改教师密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "修改教师密码",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "查询过滤器，如果没有默认查询所以信息",
                        "name": "action",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "需要修改的密码",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/teacher/student/new": {
            "post": {
                "description": "添加/注册一个学生",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "添加/注册一个学生",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新用户用户名",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "新用户邮箱",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新用户真实姓名",
                        "name": "realName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新用户学号",
                        "name": "userSchoollD",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新用户学校代码",
                        "name": "schoolCode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新用户班级",
                        "name": "class",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新用户性别",
                        "name": "sex",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/teacher/{action}": {
            "get": {
                "description": "获取教师相关信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "获取教师相关信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "查询过滤器，如果没有默认查询所以信息",
                        "name": "action",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-model_GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errhandle.ErrCode": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14
            ],
            "x-enum-varnames": [
                "NoError",
                "InnerError",
                "TokenError",
                "ParamsError",
                "UserExists",
                "UserNonExists",
                "PasswordInvalid",
                "PermissionDenied",
                "ClassError",
                "SexError",
                "NameExists",
                "EmailFormatError",
                "RealNameFormatError",
                "SchoolError",
                "PasswordTooShort"
            ]
        },
        "model.CommonResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
...
        "model.CommonResponse-model_GetUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.GetUserResponse"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
...
        "model.GetUserResponse": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "realName": {
                    "type": "string"
                },
                "schoolCode": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                },
                "userSchoollD": {
                    "type": "string"
                }
            }
        },
...
    }
}